feature(controller):signup end-of-creation-of-signup-controller [finishes #171686868]
