+++ .eslintrc.json
@@ -16,6 +16,8 @@
     },
     "rules": {
         "linebreak-style":"off",
-        "no-console":"off"
+        "no-console":"off",
+        "newline-per-chained-call":"off",
+        "arrow-parens":"off"
     }
 }
